Move warp features to map service and remove npc service
