+++ src/api/server.ts
@@ -30,9 +30,6 @@ import { createMonsterRepository } from "./services/monster/repository";
 import { mapDefinition } from "./services/map/definition";
 import { mapController } from "./services/map/controller";
 import { createMapRepository } from "./services/map/repository";
-import { createNpcRepository } from "./services/npc/repository";
-import { npcDefinition } from "./services/npc/definition";
-import { npcController } from "./services/npc/controller";
 
 const args = readCliArgs(options);
 const logger = createLogger(
@@ -63,8 +60,7 @@ const linker = createPublicFileLinker({
 
 const itemRepository = createItemRepository({ yaml, files, ...args });
 const monsterRepository = createMonsterRepository({ ...args, yaml, npc });
-const mapRepository = createMapRepository({ files, linker, formatter });
-const npcRepository = createNpcRepository(npc);
+const mapRepository = createMapRepository({ files, linker, formatter, npc });
 
 app.use(auth.middleware);
 app.use(cors());
@@ -74,7 +70,6 @@ app.use(rpc(itemDefinition, itemController(itemRepository)));
 app.use(rpc(authDefinition, authController({ db, yaml, auth, ...args })));
 app.use(rpc(monsterDefinition, monsterController(monsterRepository)));
 app.use(rpc(mapDefinition, mapController(mapRepository)));
-app.use(rpc(npcDefinition, npcController(npcRepository)));
 app.use(
   rpc(
     metaDefinition,
