+++ src/api/services/map/definition.ts
@@ -3,7 +3,13 @@ import { createTagFactory } from "../../../lib/createTagFactory";
 import { createRpcDefinition } from "../../util/rpc";
 import { UserAccessLevel } from "../auth/types";
 import { createSearchTypes } from "../search/types";
-import { mapIdType, mapInfoFilter, mapInfoType } from "./types";
+import {
+  mapIdType,
+  mapInfoFilter,
+  mapInfoType,
+  warpFilter,
+  warpType,
+} from "./types";
 
 const infoTag = createTagFactory("MapInfo");
 const imageTag = createTagFactory("MapImage");
@@ -17,6 +23,7 @@ export const mapDefinition = createRpcDefinition({
         ...createSearchTypes(mapInfoType, mapInfoFilter.type),
         { tags: (res) => infoTag.many(res?.entities.map((map) => map.id)) }
       )
+      .query("searchWarps", ...createSearchTypes(warpType, warpFilter.type))
       .query("getMap", mapIdType, mapInfoType, {
         tags: (map) => [infoTag.one(map?.id)],
       })
