+++ src/api/services/map/repository.ts
@@ -5,17 +5,26 @@ import { FileStore } from "../../../lib/createFileStore";
 import { parseLuaTableAs } from "../../common/parseLuaTableAs";
 import { Linker } from "../../../lib/createPublicFileLinker";
 import { ImageFormatter } from "../../../lib/createImageFormatter";
-import { MapId, MapInfo, MapInfoPostProcess, mapInfoType } from "./types";
+import { NpcDriver } from "../../rathena/NpcDriver";
+import {
+  MapId,
+  MapInfo,
+  MapInfoPostProcess,
+  mapInfoType,
+  warpType,
+} from "./types";
 export type MapRepository = ReturnType<typeof createMapRepository>;
 
 export function createMapRepository({
   files,
   linker,
   formatter,
+  npc,
 }: {
   files: FileStore;
   linker: Linker;
   formatter: ImageFormatter;
+  npc: NpcDriver;
 }) {
   const info: Record<string, MapInfo> = {};
 
@@ -32,6 +41,7 @@ export function createMapRepository({
     info,
     updateInfo: infoFile.update,
     countImages: () => fs.readdirSync(mapLinker.directory).length,
+    warps: npc.resolve("scripts_warps.conf", warpType),
     async hasImage(mapId: MapId) {
       return fs.promises
         .access(mapImagePath(mapId))
