--- src/api/services/npc/definition.ts
@@ -1,11 +1,0 @@
-import { createRpcDefinition } from "../../util/rpc";
-import { createSearchTypes } from "../search/types";
-import { warpFilter, warpType } from "./types";
-
-export const npcDefinition = createRpcDefinition({
-  entries: (builder) =>
-    builder.query(
-      "searchWarps",
-      ...createSearchTypes(warpType, warpFilter.type)
-    ),
-});
